Remove test from MainActivity.kt
